+++ lib/src/model/qr_code_image.dart
@@ -1,0 +1,5 @@
+class QrCodeImage {
+  String? urlCode;
+  String? base64Image;
+  QrCodeImage({this.urlCode, this.base64Image});
+}