+++ lib/whatsapp_bot_flutter.dart
@@ -1,3 +1,5 @@
+export 'package:whatsapp_bot_flutter/src/model/call_event.dart';
+export 'package:whatsapp_bot_flutter/src/whatsapp_client.dart';
 export 'package:whatsapp_bot_flutter/src/model/whatsapp_file_type.dart';
 export 'package:whatsapp_bot_flutter/src/model/message.dart';
 export 'package:whatsapp_bot_flutter/src/model/connection_event.dart';